% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_table.R
\docType{data}
\name{brain_table}
\alias{brain_table}
\title{Hi-C data from two regions of the brain at 100KB resolution}
\format{
An object of class \code{list} of length 2.
}
\usage{
brain_table
}
\value{
A hic.table list
}
\description{
An hic.table object containing data from chromosomes 18 & 22.
    hic_loess has already been used to jointly normalize the datasets.
}
\keyword{datasets}
