% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centromere_locations.R
\docType{data}
\name{centromere_locations}
\alias{centromere_locations}
\title{Locations of the centromeres for hg19}
\format{
A data.frame with 3 columns and 24 rows
}
\usage{
centromere_locations
}
\value{
A data.frame
}
\description{
A BED file
}
\keyword{datasets}
