% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooler.R
\docType{data}
\name{cooler}
\alias{cooler}
\title{Hi-C data in the cooler format}
\format{A matrix with 4011620 rows and 7 columns in BEDPE format.}
\source{
Data from the Index of Coolers. \url{ftp://cooler.csail.mit.edu/coolers/hg19/Dixon2012-H1hESC-HindIII-allreps-filtered.1000kb.cool}
}
\usage{
cooler
}
\value{
A matrix.
}
\description{
Contains Hi-C data for the entire genome from the Dixon2012-H1hESC-HindIII-allreps-filtered.1000kb.cool file.
}
\keyword{datasets}
