% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19_blacklist.R
\docType{data}
\name{hg19_blacklist}
\alias{hg19_blacklist}
\title{BED file for hg19 blacklisted regions}
\format{
A data.frame with 3 columns and 1649 rows:
    \describe{
    \item{chr}{chromosome for the region}
    \item{start}{start location of the region}
    \item{end}{end location of the region}
 }
}
\source{
Data from UCSC
    \url{http://genome.ucsc.edu/cgi-bin/hgFileUi?db=hg19&g=wgEncodeMapability}
}
\usage{
hg19_blacklist
}
\value{
A data.frame
}
\description{
A BED file with regions which may be
    excluded from your Hi-C data analysis.
}
\keyword{datasets}
