% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38_blacklist.R
\docType{data}
\name{hg38_blacklist}
\alias{hg38_blacklist}
\title{BED file for hg38 blacklisted regions}
\format{
A data.frame with 3 columns and 38 rows:
    \describe{
    \item{chr}{chromosome for the region}
    \item{start}{start location of the region}
    \item{end}{end location of the region}
 }
}
\source{
Data from
    \url{http://mitra.stanford.edu/kundaje/akundaje/release/blacklists/hg38-human/}
}
\usage{
hg38_blacklist
}
\value{
A data.frame
}
\description{
A BED file with regions which may be
    excluded from your Hi-C data analysis.
}
\keyword{datasets}
