% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hic_compare.R
\name{hic_compare}
\alias{hic_compare}
\title{Detect differences between two jointly normalized Hi-C datasets.}
\usage{
hic_compare(hic.table, Plot = FALSE, Plot.smooth = TRUE, parallel = FALSE,
  BP_param = bpparam())
}
\arguments{
\item{hic.table}{A hic.table or list of hic.tables output from the
\code{hic_loess} function. hic.table must be jointly normalized
before being entered.}

\item{Plot}{Logical, should the MD plot showing before/after loess normalization
be output?}

\item{Plot.smooth}{Logical, defaults to TRUE indicating the MD plot
will be a smooth scatter plot. Set to FALSE for a scatter plot
with discrete points.}

\item{parallel}{Logical, set to TRUE to utilize the \code{parallel} package's
parallelized computing. Only works on unix operating systems. Only useful if
entering a list of hic.tables.}

\item{BP_param}{Parameters for BiocParallel. Defaults to bpparam(), see help
for BiocParallel for more information
\url{http://bioconductor.org/packages/release/bioc/vignettes/BiocParallel/
inst/doc/Introduction_To_BiocParallel.pdf}}
}
\value{
A hic.table with additional columns containing a p-value for the significance
    of the difference and the raw fold change between the IFs of the two datasets.
}
\description{
Detect differences between two jointly normalized Hi-C datasets.
}
\details{
The function takes in a hic.table or a list of hic.table objects created
    with the \code{hic_loess} function. If you wish to perform difference
    detection on Hi-C data for multiple chromosomes use a list of hic.tables. The process
    can be parallelized using the \code{parallel}
    setting. The adjusted IF and adjusted M calculated from \code{hic_loess} are used for
    difference detection. Fisher's exact test is performed for each pair of IFs between
    the two datasets. For example if IF1 = 50 and IF2 = 100 we create a 2x2 table of the
    form \code{matrix(c(50, 100, 75, 75), ncol=2)}. The null hypothesis is that IF1 = IF2
    = Average expression of IF1 and IF2. We can then perform fisher's exact test on this
    2x2 table. FDR p-value adjustment for multiple testing is then performed on a per
    distance basis. i.e. at each distance the vector of p-values corresponding to the
    interactions occuring at that distance have the FDR multiple testing correction 
    applied to them. See methods of Stansfield & Dozmorov 2017 for more details.
}
\examples{
# Create hic.table object using included Hi-C data in sparse upper triangular
# matrix format
data('HMEC.chr22')
data('NHEK.chr22')
hic.table <- create.hic.table(HMEC.chr22, NHEK.chr22, chr = 'chr22')
# Plug hic.table into hic_loess()
result <- hic_loess(hic.table, Plot = TRUE)
# perform difference detection
diff.result <- hic_compare(result, Plot = TRUE)

}
