% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCdiff.r
\docType{package}
\name{HiCdiff-package}
\alias{HiCdiff}
\alias{HiCdiff-package}
\title{HiCdiff}
\description{
HiCdiff provides functions for joint normalization and difference detection in multiple Hi-C datasets.
HiCdiff operates on processed Hi-C data in the form of chromosome-specific chromatin interaction matrices.
It accepts three-column tab-separated text files storing chromatin interaction matrices in a sparse matrix format
which are available from several sources. HiCdiff is designed to give the user the ability to perform a comparative
analysis on the 3-Dimensional structure of the genomes of cells in different biological states.
}
\details{
To learn more about HiCdiff, start with the vignettes:
`browseVignettes(package = "HiCdiff")`
}
\author{
\strong{Maintainer}: John Stansfield \email{stansfieldjc@vcu.edu}

Authors:
\itemize{
  \item Mikhail Dozmorov \email{mikhail.dozmorov@vcuhealth.org}
}

}
