% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NHEK.chr10.R
\docType{data}
\name{NHEK.chr10}
\alias{NHEK.chr10}
\title{Hi-C data from NHEK cell line - chromosome 10 at 500kb resolution}
\format{A matrix with 35510 rows and 3 columns:
    \describe{
    \item{region1}{The first interacting region - corresponds to the row
         name of a full Hi-C contact matrix}
    \item{region2}{The second interacting region - corresponds to the
         column name of a full Hi-C contact matrix}
    \item{IF}{The Interaction Frequency - number of read counts found
         for the interaction between region1 and region2}
 }}
\source{
Data from the Aiden Lab. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}
}
\usage{
NHEK.chr10
}
\value{
A matrix
}
\description{
A sparse upper triangular matrix containing the interacting regions and the
    corresponding Interaction Frequency (IF).
}
\keyword{datasets}
