% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhek.IS.R
\docType{data}
\name{nhek.IS}
\alias{nhek.IS}
\title{Hi-C data from NHEK cell line - chromosome 22 at 500kb resolution}
\format{An InteractionSet with 2508 rows and 1 metadata column:
    \describe{
    \item{IF}{The Interaction Frequency - number of read counts found
         for the interaction between region1 and region2}
 }}
\source{
Data from the Aiden Lab. See their website at
    \url{http://www.aidenlab.org/}
     Or the the GEO link to download the data
    \url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE63525}
}
\usage{
nhek.IS
}
\value{
A matrix
}
\description{
An InteractionSet object 2508 interactions and 1 metadata column
    containing the Interaction Frequencies
}
\keyword{datasets}
