% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_A.R
\name{plot_A}
\alias{plot_A}
\title{function to plot distribution of A to get input values for weight_M}
\usage{
plot_A(hic.table)
}
\arguments{
\item{hic.table}{A hic.table or list of hic.tables. You must run 
hic_loess on the hic.table before running plot_A.}
}
\value{
Plots of A.
}
\description{
function to plot distribution of A to get input values for weight_M
}
\details{
This function will plot 4 veiws of the distribution of
    A for use in determining the A.min and quant parameters for
    the weight_M function. The first plot is the densities of A
    for each chromosome cut off at the 75th quantile. The second
    plot is the densities of each chromosome cut at the 90th
    quantile. The third plot is the log2 histogram of the 
    distribution of A. The fourth plot is the histogram of
    A cut at the 75th quantile. You should try to choose a
    value of A.min where the distribution starts to only
    descend after that point. You should choose the value
    for quant at the point where the density flattens
    out at the lowest values.
}
\examples{
# Create hic.table object using included Hi-C data in sparse upper
# triangular matrix format
data("HMEC.chr22")
data("NHEK.chr22")
hic.table <- create.hic.table(HMEC.chr22, NHEK.chr22, chr= 'chr22')
# Plug hic.table into hic_loess()
hic.table <- hic_loess(hic.table, Plot = TRUE)
# plot_A
plot_A(hic.table)

}
