% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_centromere.R
\name{split_centromere}
\alias{split_centromere}
\title{Function to split hic.table into 2 subsets at the centromere}
\usage{
split_centromere(hic.table)
}
\arguments{
\item{hic.table}{A hic.table or list of hic.table objects}
}
\value{
A list of hic.tables in which the matrices have been split at the centromere
}
\description{
Function to split hic.table into 2 subsets at the centromere
}
\examples{
data('HMEC.chr22')
data('NHEK.chr22')
hic.table <- create.hic.table(HMEC.chr22, NHEK.chr22, chr = 'chr22')
split <- split_centromere(hic.table)
}
