% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pvals.R
\name{visualize_differences}
\alias{visualize_differences}
\title{Alternate heatmap for dealing with rankings}
\usage{
visualize_differences(hic.table, which_rank = "rnkMean",
  only_toprank = TRUE, proportion = 0.05)
}
\arguments{
\item{which_rank}{The column name for the ranking that you
want to plot.}

\item{only_toprank}{Logical, Should only the top ranks be 
plotted? If TRUE then the proportion option will be used
to determine how many ranks will be considered top.}

\item{proportion}{The proportion of ranks to be considered
as top ranks. Defaults to 0.05.}
}
\description{
Alternate heatmap for dealing with rankings
}
