% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_pvals.R
\name{visualize_pvals}
\alias{visualize_pvals}
\title{Function to visualize p-values from HiCcompare results}
\usage{
visualize_pvals(hic.table)
}
\arguments{
\item{hic.table}{A hic.table object that has been
normalized and has had differences detected.}
}
\value{
A heatmap
}
\description{
Function to visualize p-values from HiCcompare results
}
\details{
The goal of this function is to visualize
    where in the Hi-C matrix the differences are
    occuring between two experimental conditions.
    The function will produce a heatmap of the
    -log10(p-values) * sign(adj.M) 
    to visualize where the
    significant differences between the datasets
    are occuring on the genome.
}
\examples{
# Create hic.table object using included Hi-C data in sparse upper triangular
# matrix format
data('HMEC.chr22')
data('NHEK.chr22')
hic.table <- create.hic.table(HMEC.chr22, NHEK.chr22, chr = 'chr22')
# Plug hic.table into hic_loess()
result <- hic_loess(hic.table, Plot = TRUE)
# perform difference detection
diff.result <- hic_diff(result, diff.thresh = 'auto', Plot = TRUE)
# visualize p-values
visualize_pvals(diff.result)
}
