% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_M.R
\name{weight_M}
\alias{weight_M}
\title{Function to weight M values on the MD plot by Average expression}
\usage{
weight_M(hic.table, A.min = 1, quant = 0.75, Plot.diagnostic = TRUE,
  Plot.MD = TRUE)
}
\arguments{
\item{hic.table}{A hic.table object or hic.table list. 
You must run hic_loess on the hic.table before performing M weighting.}

\item{A.min}{The A value at which to start the powerlaw estimation of the
density. This should be determined by looking at the plots of A from
the plot_A function. See vignette for more details. Defaults to 1.}

\item{quant}{The quantile value at which to stop weighting M values.
Typical values should range from 0.6 to 0.9. This can be determined
by looking at the plots of the distribution of A from the plot_A
function. See vignette for more details. Defaults to 0.75.}

\item{Plot.diagnostic}{Logical, should a histogram of the subset of 
A between A.min and A < quant be plotted with the powerlaw fit
overlaid. It is recommend to check the fit of the powerlaw to 
the distribution. If a good fit is not achieved you may need
to adjust A.min or quant parameters. Defaults to TRUE.}

\item{Plot.MD}{Logical, should the MD plot of the weighted M 
values be plotted?  Defaults to TRUE.}
}
\value{
A hic.table with weighted adj.M values.
}
\description{
Function to weight M values on the MD plot by Average expression
}
\details{
Weighting of M is used as a way to account for the fact
    that IFs with low average values are less trustworthy compared
    to IFs with high average values. i.e. a comparison between 
    IF1 = 1 and IF2 = 2 results in an M value of 1, but a 
    comparison of IF1 = 100 and IF2 = 200 also resuts in an
    M value of 1. However the average expression of the first
    comparison is 1.5 while the second is 150. Naturally the
    first comparison is much more likely to have occured as a 
    fluke compared to a true biological difference between the
    conditions being tested. This function performs down-weighting
    of M values which arise from IFs exhibiting low average
    expression. For a more in-depth explanation with examples
    please see the vignette.
}
\examples{
# Create hic.table object using included Hi-C data in sparse upper
# triangular matrix format
data("HMEC.chr22")
data("NHEK.chr22")
hic.table <- create.hic.table(HMEC.chr22, NHEK.chr22, chr= 'chr22')
# Plug hic.table into hic_loess()
hic.table <- hic_loess(hic.table, Plot = TRUE)
# weight M-values
hic.table <- weight_M(hic.table, A.min = 75, quant = 0.9)

}
