% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCcompare.r
\docType{package}
\name{HiCcompare-package}
\alias{HiCcompare}
\alias{HiCcompare-package}
\title{HiCcompare}
\description{
HiCcompare provides functions for joint normalization and difference detection in multiple Hi-C datasets.
HiCcompare operates on processed Hi-C data in the form of chromosome-specific chromatin interaction matrices.
It accepts three-column tab-separated text files storing chromatin interaction matrices in a sparse matrix format
which are available from several sources. HiCcompare is designed to give the user the ability to perform a comparative
analysis on the 3-Dimensional structure of the genomes of cells in different biological states.
}
\details{
To learn more about HiCcompare, start with the vignettes:
`browseVignettes(package = "HiCcompare")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dozmorovlab/HiCcompare}
  \item Report bugs at \url{https://github.com/dozmorovlab/HiCcompare/issues}
}

}
\author{
\strong{Maintainer}: Mikhail Dozmorov \email{mikhail.dozmorov@gmail.com} (\href{https://orcid.org/0000-0002-0086-8358}{ORCID})

Authors:
\itemize{
  \item Kellen Cresswell \email{cresswellkg@vcu.edu}
  \item John Stansfield \email{stansfieldjc@vcu.edu}
}

}
