% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCcompare.r, R/HiCdiff.R
\docType{package}
\name{HiCcompare-package}
\alias{HiCcompare}
\alias{HiCcompare-package}
\alias{HiCcompare}
\alias{HiCcompare-package}
\title{HiCcompare}
\description{
HiCcompare provides functions for joint normalization and difference detection in multiple Hi-C datasets.
HiCcompare operates on processed Hi-C data in the form of chromosome-specific chromatin interaction matrices.
It accepts three-column tab-separated text files storing chromatin interaction matrices in a sparse matrix format
which are available from several sources. HiCcompare is designed to give the user the ability to perform a comparative
analysis on the 3-Dimensional structure of the genomes of cells in different biological states.

HiCdiff provides functions for joint normalization and difference detection in multiple Hi-C datasets.
HiCdiff operates on processed Hi-C data in the form of chromosome-specific chromatin interaction matrices.
It accepts three-column tab-separated text files storing chromatin interaction matrices in a sparse matrix format
which are available from several sources. HiCdiff is designed to give the user the ability to perform a comparative
analysis on the 3-Dimensional structure of the genomes of cells in different biological states.
}
\details{
To learn more about HiCcompare, start with the vignettes:
`browseVignettes(package = "HiCcompare")`

To learn more about HiCdiff, start with the vignettes:
`browseVignettes(package = "HiCdiff")`
}
\author{
\strong{Maintainer}: John Stansfield \email{stansfieldjc@vcu.edu}

Authors:
\itemize{
  \item Mikhail Dozmorov \email{mikhail.dozmorov@vcuhealth.org}
}


\strong{Maintainer}: John Stansfield \email{stansfieldjc@vcu.edu}

Authors:
\itemize{
  \item Mikhail Dozmorov \email{mikhail.dozmorov@vcuhealth.org}
}

}
